backend/apps/categories/views.py
from rest_framework.viewsets import(
    ModelViewSet,
    ReadOnlyModelViewSet,
    GenericViewSet,
)

from rest_framework.views import(
    APIView,
)

from rest_framework.mixins import(
    CreateModelMixin,
    UpdateModelMixin,
    DestroyModelMixin,
    ListModelMixin,
    RetrieveModelMixin,
)

from rest_framework.response import Response

from rest_framework.permissions import(
    IsAuthenticated,
    IsAuthenticatedOrReadOnly,
)

from django_filters.rest_framework import(
    DjangoFilterBackend,
)

from rest_framework import filters

from rest_framework.decorators import action

